chatbot/app/requirements.txt
ipython
google-cloud
google-cloud-storage==2.14.0
fsspec==2024.2.0
gcsfs==2024.2.0
google-cloud-pubsub
python-dotenv
Flask-MySQLdb
Flask-SQLAlchemy
pymysql
python-jose 
flask-jwt-extended
fastapi
bcrypt
passlib
phpserialize
cryptography
Flask
google-generativeai
flask-cors
Flask-Limiter
flask-swagger-ui
gunicorn